Break BeaconChainHarness validator gen into fn
Prepares for allowing for loading from file
